requirements.txt
matplotlib
librosa
soundfile
numpy
absl-py
tensorflow